+++ pkg/api/api.go
@@ -33,7 +33,6 @@ import (
 	"github.com/cortexproject/cortex/pkg/storegateway"
 	"github.com/cortexproject/cortex/pkg/storegateway/storegatewaypb"
 	"github.com/cortexproject/cortex/pkg/util/push"
-	"github.com/cortexproject/cortex/pkg/util/runtimeconfig"
 )
 
 type Config struct {
@@ -179,10 +178,11 @@ func (a *API) RegisterAPI(httpPathPrefix string, actualCfg interface{}, defaultC
 }
 
 // RegisterRuntimeConfig registers the endpoints associates with the runtime configuration
-func (a *API) RegisterRuntimeConfig(runtimeCfgManager *runtimeconfig.Manager) {
+func (a *API) RegisterRuntimeConfig(runtimeConfigHandler http.HandlerFunc) {
 	a.indexPage.AddLink(SectionAdminEndpoints, "/runtime_config", "Current Runtime Config (incl. Overrides)")
+	a.indexPage.AddLink(SectionAdminEndpoints, "/runtime_config?mode=diff", "Current Runtime Config (show only values that differ from the defaults)")
 
-	a.RegisterRoute("/runtime_config", runtimeConfigHandler(runtimeCfgManager), false, "GET")
+	a.RegisterRoute("/runtime_config", runtimeConfigHandler, false, "GET")
 }
 
 // RegisterDistributor registers the endpoints associated with the distributor.